Removed the TTS/STT changes
