+++ src/jarvis/jarvis/engines/__init__.py
@@ -1,4 +1,3 @@
-from .stt_google import *
-from .stt_vosk import *
+from .stt import *
 from .tts import *
 from .ttt import *
